Refactor reporting config helpers
