+++ app/reporting.py
@@ -12,6 +12,8 @@ from pptx import Presentation
 from pptx.util import Inches
 import os
 
+from .config import _get_api_key, _get_model, _read_config
+
 
 class Summary(BaseModel):
     title: str
@@ -33,35 +35,6 @@ class SlideDeck(BaseModel):
     slides: List[Slide]
 
 
-# Configuration helpers redefined here to avoid circular imports
-CONFIG_FILE = os.path.join(os.path.dirname(__file__), "config.json")
-
-
-def _read_config() -> dict:
-    conf = {}
-    if os.path.exists(CONFIG_FILE):
-        with open(CONFIG_FILE, "r") as f:
-            conf.update(json.load(f))
-    env_key = os.getenv("OPENROUTER_API_KEY")
-    if env_key:
-        conf.setdefault("openrouter_api_key", env_key)
-    env_model = os.getenv("OPENROUTER_MODEL")
-    if env_model:
-        conf.setdefault("openrouter_model", env_model)
-    return conf
-
-
-def _get_api_key() -> str:
-    conf = _read_config()
-    key = conf.get("openrouter_api_key")
-    if not key:
-        raise RuntimeError("OpenRouter API key not configured")
-    return key
-
-
-def _get_model() -> str:
-    conf = _read_config()
-    return conf.get("openrouter_model", "openai/gpt-3.5-turbo")
 
 
 # LLM generation functions
